1.5.5 update
* Added support for asset decryption via the CzDecrypt class, decrypts
all xml and lua assets before running them
* If actor has geometry assigned and no shape and it is under control of
physics then the geometry will be assigned as the shape as long as the
geometry is polygon based
* Added access to Matrices from Lua using the new matrix library
* Added actor Lua library functions setVertex(), setRGBA() and setUV()
which can be used to modify the actors underlying geometry
* Added support for decrypting lua and xoml assets, add decrypt tag to
appeasy.xml with attribute enable="true" to enable
* Added new unified Ads platform as well as access from Lua. Support for
Revmob and Flurry App Spot
* Fixed crash on exit that was causing pause
* Creating XOML from a template no longer requires a container, if one
not supplied then assets will be placed in global resource space
* Fixed docked actors that are scaled
* Fixed issue in where proportional sizing was not taking into account
virtual resolution scaling
* Added more actor / scene properties and rearranged by common use to
improve performance

+++ source/Script/lua/CzLuaSys.cpp
@@ -492,7 +492,7 @@ static int LUA_setBackgroundColour(lua_State *lua)
 	// Get colour cponents
 	int red = 0;
 	if (lua_isnumber(lua, 1))
-		red = lua_tonumber(lua, 1);
+		red = (int)lua_tonumber(lua, 1);
 	else
 	{
 		CzScriptEngineLua::DisplayError(lua, "sys.setBackgroundColour() invalid value (Param0), expected number");
@@ -501,7 +501,7 @@ static int LUA_setBackgroundColour(lua_State *lua)
 	}
 	int green = 0;
 	if (lua_isnumber(lua, 2))
-		green = lua_tonumber(lua, 2);
+		green = (int)lua_tonumber(lua, 2);
 	else
 	{
 		CzScriptEngineLua::DisplayError(lua, "sys.setBackgroundColour() invalid value (Param1), expected number");
@@ -510,7 +510,7 @@ static int LUA_setBackgroundColour(lua_State *lua)
 	}
 	int blue = 0;
 	if (lua_isnumber(lua, 3))
-		blue = lua_tonumber(lua, 3);
+		blue = (int)lua_tonumber(lua, 3);
 	else
 	{
 		CzScriptEngineLua::DisplayError(lua, "sys.setBackgroundColour() invalid value (Param2), expected number");
